requirements.txt
tensorflow-macos==2.9.0
pandas==1.4.2
numpy==1.22.4
scipy==1.7.3
scikit-learn
jupyter
matplotlib
kaggle
spacy
nltk
enchant
tensorflow-addons
transformers
seaborn
wordcloud
unidecode
gensim
keras_contrib
torch